Move field finding to traverse package
