+++ src/test/java/no/ssb/dapla/dataset/doc/traverse/FieldFinderTest.java
@@ -1,4 +1,4 @@
-package no.ssb.dapla.dataset.doc.model.lineage;
+package no.ssb.dapla.dataset.doc.traverse;
 
 import no.ssb.dapla.dataset.doc.template.TestUtils;
 import org.apache.avro.Schema;
